Remove most default settings
Most of them were only provided as targets for inline overrides,
but those have been removed for SL4. The user may specify these in the
`args` setting directly.

Also remove some hard-coded but unnecessary default args, i.e.
`--ignore-missing-imports` and `--follow-imports`. We want the linting
experience to be as close to the default as possible, unless the user
specifically overrides them.

Fixes #19.