Only check mypy version once per lint
Previously it would get called for *each* lint result. Now we just check
at the start of the lint.

Note that Linter instances are shared for the same view, but this
doesn't matter here since the same view will also run the same
executable, so it's safe to override the instance variable.