Allow only one process per directory
mypy actually operates with caches (probably even using sqlite) and these
aren't locked while running. This leads to various errors when we try to
run it concurrently.

SublimeLinter itself has no flag which forbids concurrent processing, so
we just lock it ourself. Creating and acquiring a lock using a
`defaultdict` is in itself not thread aware but it is simple enough here.