Allow setting `cache-dir` to `False` to omit the arg
The user can set `cache-dir` to `False` to omit the arg from the command
completely. This is especially useful if the actual cache dir is configured
in the `mypy.ini` file.

Omit automatically in case the default location `.mypy_cache` exists.
Likely it is populated and the right thing.

In all cases do not manually edit `cmd`. Just set the setting and let
SublimeLinter do the rest.