Support "--show-error-end" from newer mypy's
We can't enable this here for everyone because it has been just
introduced in v0.981 (Sep 2022).  But we can support its output format.

Per spec, mypy might report `-1`s if it doesn't know a better value.
We convert them to `None`s as that is what SL expects.  Except for
`line` which is mandatory in SL land and hereby reported on the first
line of the buffer.